Fixed bug where i forgot weekends
