+++ Announcer/Support Files/Delegates/SceneDelegate.swift
@@ -60,4 +60,3 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         continueFromCoreSpotlight(with: userActivity)
     }
 }
-
